+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -23,6 +23,7 @@ public class Http11Processor implements Runnable, Processor {
             new CssRequestHandler(),
             new JsRequestHandler(),
             new LoginRequestHandler(),
+            new LoginPostRequestHandler(),
             new RegisterGetRequestHandler(),
             new RegisterPostRequestHandler()
     );
